requirements.txt
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
binance-futures-connector==3.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
requests==2.28.1
service-identity==21.1.0
six==1.16.0
tradingview-ta==3.3.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.13
zope.interface==5.5.2
